test(skills): split structure lint coverage
Separate rule-isolation coverage from declaration-removal and CLI behavior while keeping a small compatibility marker and shared fixture support. Preserve the exact 25 test-name set and all exit-code assertions.

Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)

Constraint: Preserve offline structure-lint behavior and every existing assertion; change test organization only.
Rejected: SIZE_OK exception | the existing classes provide clean responsibility seams below 250 pure LOC.
Confidence: high
Scope-risk: narrow
Directive: Keep future structure-lint test modules below 250 pure LOC and split by rule versus CLI responsibility.
Tested: 37 focused tests; CLI exit 0/1/2 cases; test-name equality; Ruff; compileall; pure-LOC gate; git diff --check.
Not-tested: Full discovery and remote CI after this test-only split.
Plan: .omo/plans/omh-domain-skill-quality-upgrades.md
